refactor: Move welcome ui and vm to onboarding
