+++ app/src/main/java/to/bitkit/ui/onboarding/WelcomeViewModel.kt
@@ -1,4 +1,4 @@
-package to.bitkit.ui
+package to.bitkit.ui.onboarding
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
